+++ dispatch/BESFSDir.cc
@@ -146,7 +146,7 @@ BESFSDir::loadDir()
 			int match_ret = reg_expr.match( dirEntry.c_str(),
 						        dirEntry.length() ) ;
 			cerr << "match returned " << match_ret << endl ;
-			if( match_ret == dirEntry.length() )
+			if( match_ret == static_cast<int>(dirEntry.length()) )
 		        {
                             _fileList.push_back(BESFSFile(_dirName, dirEntry));
                         }
